fix : entity 및 Repository 수정

+++ bank/backend/src/main/java/com/bank/backend/entity/BankCode.java
@@ -14,7 +14,7 @@ public class BankCode {
 
     @Id
     @Column(name="bankCodeId")
-    private int bankCodeId;
+    private Long bankCodeId;
 
     @Column(name="bankName", nullable = false, length = 20)
     private String bankName;
